+++ src/chart/types.rs
@@ -25,6 +25,8 @@ pub struct MinimalChartData {
     pub usd_value: f64,
 }
 
+#[allow(dead_code)]
+// TODO: use this
 #[derive(serde::Serialize)]
 #[serde(untagged)]
 pub enum ChartResponse {
