+++ src/coingecko.rs
@@ -4,10 +4,12 @@ use reqwest::Client;
 use serde::Deserialize;
 use serde_json::{from_reader, Value};
 
-use crate::{chart::types::GetCoinMarketChartParams, Error};
+use crate::{chart::types::GetCoinMarketChartParams, enums::Error};
 
 const BASE_URL: &str = "https://api.coingecko.com/api/v3";
 
+#[allow(dead_code)]
+// TODO: avoid dead code
 #[derive(Deserialize, Debug)]
 pub struct CoingeckoToken {
     pub name: String,
