+++ src/consts.rs
@@ -1,0 +1,4 @@
+pub const USDC_ADDRESS: &str = "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v";
+pub const SOL_ADDRESS: &str = "So11111111111111111111111111111111111111112";
+pub const USDC_IMAGE: &str = "https://s2.coinmarketcap.com/static/img/coins/128x128/3408.png";
+pub const SOL_IMAGE: &str = "https://s2.coinmarketcap.com/static/img/coins/128x128/5426.png";
