+++ src/enums.rs
@@ -1,0 +1,9 @@
+#[derive(Debug)]
+pub enum Error {
+    InvalidAddress,
+    InvalidIndex,
+    InvalidTimeframe,
+    FetchError,
+    ParseError,
+    TooManyRequests,
+}
